+++ tmp/build-release/src/CMakeFiles/oatpp-mariadb.dir/cmake_clean.cmake
@@ -1,0 +1,28 @@
+file(REMOVE_RECURSE
+  "CMakeFiles/oatpp-mariadb.dir/link.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/Connection.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/Connection.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ConnectionProvider.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ConnectionProvider.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/Executor.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/Executor.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/QueryResult.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/QueryResult.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/Deserializer.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/Deserializer.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/ResultMapper.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/ResultMapper.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/Serializer.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/mapping/Serializer.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ql_template/Parser.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ql_template/Parser.cpp.o.d"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ql_template/TemplateValueProvider.cpp.o"
+  "CMakeFiles/oatpp-mariadb.dir/oatpp-mariadb/ql_template/TemplateValueProvider.cpp.o.d"
+  "liboatpp-mariadb.pdb"
+  "liboatpp-mariadb.so"
+)
+
+# Per-language clean rules from dependency scanning.
+foreach(lang CXX)
+  include(CMakeFiles/oatpp-mariadb.dir/cmake_clean_${lang}.cmake OPTIONAL)
+endforeach()