+++ tmp/build-release/test/CTestTestfile.cmake
@@ -1,0 +1,8 @@
+# CMake generated Testfile for 
+# Source directory: /home/tiger/CascadeProjects/oatpp-mariadb/test
+# Build directory: /home/tiger/CascadeProjects/oatpp-mariadb/tmp/build-release/test
+# 
+# This file includes the relevant testing commands required for 
+# testing this directory and lists subdirectories to be tested as well.
+add_test(oatpp-mariadb-tests "oatpp-mariadb-tests")
+set_tests_properties(oatpp-mariadb-tests PROPERTIES  _BACKTRACE_TRIPLES "/home/tiger/CascadeProjects/oatpp-mariadb/test/CMakeLists.txt;37;add_test;/home/tiger/CascadeProjects/oatpp-mariadb/test/CMakeLists.txt;0;")